Isolate config header utility from PyRAF
