+++ docs/refactor_plan.md
@@ -39,22 +39,40 @@ outputs or explicit before/after comparisons.
   - HIRES-J / HIRES-Y input-list parsing
   - required aperture fields when manual aperture mode is enabled
   - `TEST/WIDE/paramSample.txt` parameter parsing
+  - `warp.config` import does not import PyRAF
 - Updated science smoke-test shell scripts.
   - default to `python3`
   - allow selecting a Python executable, for example
     `PYTHON=python3.11 ./testWarpSci.sh`
   - stop on shell errors with `set -eu`
 - Documented test commands in README.
+- Isolated a small FITS-header helper from PyRAF-dependent code.
+  - Moved `header_key_read` to `warp/fits_utils.py`.
+  - Kept `warp.Spec2Dtools.header_key_read` available for compatibility.
+  - `warp.config` can now be imported without importing PyRAF.
+- Verified lightweight tests on newer Python versions.
+  - Python 3.7.16 -> 7 passed
+  - Python 3.12.11 -> 7 passed
+  - Python 3.13.5 -> 7 passed
+- Verified one full WIDE fast smoke run with the current Astroconda/PyRAF
+  environment.
+  - Explicit environment:
+    `IRAFARCH=macos64 iraf=/Users/hamano/iraf/iraf-2.17.1/`
+  - Command shape:
+    `python Warp_sci.py ./TEST/WIDE/4_Ari_list.txt ... -d /private/tmp/warp_codex_4_Ari_WIDE_test -f`
+  - Result: exited successfully and printed `=== Finished. ===`
 
 ## Related Commits
 
 - `83d9749 Add gitignore for local artifacts`
 - `6623e6f Remove tracked local artifacts`
 - `60ce277 Add lightweight input parsing tests`
+- `a52abd8 Document refactor plan`
+- `1f9e977 Document PyRAF environment caveat`
 
 ## Next Steps
 
-1. Push the current low-risk maintenance/test commits.
+1. Commit and push the current PyRAF import-isolation change.
 2. Add a small environment definition.
    - Start by recording the currently working environment.
    - Treat newer Python targets as a separate follow-up.
@@ -66,10 +84,10 @@ outputs or explicit before/after comparisons.
    - Compare important FITS headers.
    - Compare array shapes and simple statistics.
    - Later, compare selected numerical outputs with tolerances.
-5. Modernize Python compatibility in small patches.
-   - Identify PyRAF side effects that happen at import time.
+5. Continue modernizing Python compatibility in small patches.
+   - Identify remaining PyRAF side effects that happen at import time.
    - Separate IRAF-dependent processing from pure parsing/utilities.
-   - Run lightweight tests on newer Python versions.
+   - Run lightweight tests on newer Python versions after each change.
 6. Once test coverage is stronger, start behavior-preserving refactors.
    - Clean up import paths.
    - Clean up configuration parsing.
@@ -90,15 +108,15 @@ outputs or explicit before/after comparisons.
   When running PyRAF/WARP tests from non-interactive tools, explicitly set:
   - `IRAFARCH=macos64`
   - `iraf=/Users/hamano/iraf/iraf-2.17.1/`
-- `warp.config` imports `warp.Spec2Dtools`, and `warp.Spec2Dtools` imports
-  PyRAF at module import time.
-  Because of this, even simple config tests can depend on a working IRAF
-  installation.
-  The newly added pytest file temporarily stubs `pyraf`.
-  A better long-term fix is to move `header_key_read` or make PyRAF imports
-  lazy.
+- `warp.config` previously imported `warp.Spec2Dtools`, and
+  `warp.Spec2Dtools` imports PyRAF at module import time.
+  This made simple config tests depend on a working IRAF installation.
+  `header_key_read` has now been moved to `warp/fits_utils.py`, so
+  `warp.config` no longer imports PyRAF.
+  There may still be other modules with PyRAF import-time side effects.
 - `testWarpSci*.sh` still runs the real pipeline.
   It may fail if output directories already exist.
   That behavior is unchanged.
-- The first lightweight tests were verified with Python 3.7.16.
-  - `python3 -m pytest -q` -> 6 passed
+- The first lightweight tests were verified with Python 3.7.16, 3.12.11, and
+  3.13.5.
+  - `python3 -m pytest -q` -> 7 passed
